feat: keep screen alive convenience for android
